refactor: Move operation under segment_list
